avatar-creation-app/backend/requirements.txt
--extra-index-url https://pypi.org/simple
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.3
python-dotenv==1.0.1
pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
cycler==0.12.1
contourpy==1.2.1
fonttools==4.51.0
kiwisolver==1.4.5
packaging==24.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==69.5.1
six==1.16.0
tzdata==2024.1
wheel==0.43.0
google-generativeai==0.3.2
httpx
